RaceAdapter done

+++ app/src/main/java/com/coldrosemob/corridas/Race.kt
@@ -1,7 +1,7 @@
 package com.coldrosemob.corridas
 
 class Race(
-    val vehicle : String,
-    val price : Float,
-    val date : String
+    val vehicle: String,
+    val price: Double,
+    val date: String
 )
